anti-plagiarism-ai/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic[dotenv]==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
websockets==12.0
redis==5.0.1
pymongo==4.6.0
bcrypt==4.1.2
python-multipart==0.0.6
aiofiles==23.2.1